Extract hodlr_algebra.h
Extracts hodlr_algebra.c functions from hodlr.h into hodlr_algebra.h
