+++ include/hmat_lib/hodlr.h
@@ -262,12 +262,3 @@ struct TreeHODLR {
   struct HODLRLeafNode *memory_leaf_ptr;
 };
 
-
-int multiply_hodlr_hodlr(
-  const struct TreeHODLR *hodlr1,
-  const struct TreeHODLR *hodlr2,
-  struct TreeHODLR *out,
-  const double svd_threshold,
-  int *ierr
-);
-
