+++ include/hmat_lib/hodlr_algebra.h
@@ -1,0 +1,11 @@
+#pragma once
+
+
+int multiply_hodlr_hodlr(
+  const struct TreeHODLR *hodlr1,
+  const struct TreeHODLR *hodlr2,
+  struct TreeHODLR *out,
+  const double svd_threshold,
+  int *ierr
+);
+
