+++ src/hodlr_algebra.c
@@ -5,6 +5,8 @@
 #include "../include/hmat_lib/error.h"
 #include "../include/hmat_lib/utils.h"
 #include "../include/hmat_lib/dense_algebra.h"
+#include "../include/hmat_lib/hodlr_algebra.h"
+
 #include "../include/internal/blas_wrapper.h"
 #include "../include/internal/lapack_wrapper.h"
